Fix V code generator to produce valid V code
- Use math.inf(1), math.inf(-1), math.nan() function-call forms instead
  of the constant references that v run rejects
- Add i64() wrapping for integers outside V's 32-bit int range
- Add INTERFACE heterogeneous strategy: emits `interface IVal {}` and
  wraps mixed-type, null-containing, and empty collection values as
  IVal(...) so v run accepts them
- Change null literal to `unsafe { nil }` so standalone null and null
  inside IVal contexts both compile
- Use map[string]IVal{} and []IVal{} for empty collections
- Regenerate all affected V golden fixtures
- Fix type annotation on `own` in _v_collect_ids_needing_wrap to resolve
  basedpyright partial-unknown diagnostics

+++ tests/integration/cases/scalar_int_large/V.v
@@ -1,5 +1,5 @@
 
 fn main() {
-	my_data := 2147483648
+	my_data := i64(2147483648)
 	_ = my_data
 }
