+++ tests/integration/cases/scalar_int_large/V_declaration_style_mut.v
@@ -1,5 +1,5 @@
 
 fn main() {
-	mut my_data := 2147483648
+	mut my_data := i64(2147483648)
 	_ = my_data
 }
