+++ tests/integration/cases/scalar_int_negative_large/V_combined.v
@@ -1,6 +1,6 @@
 
 fn main() {
-	mut my_data := -2147483649
-	my_data = -2147483649
+	mut my_data := i64(-2147483649)
+	my_data = i64(-2147483649)
 	_ = my_data
 }
